feat: split code in multiple files
